continuing "kernel then heaviside" filter Make OC optimizer algorithm aware of "Update Problem" Stage Register updating problem example fix typo caught by tests elsewhere

+++ base/src/tools/Plato_OptimizerEngineStageData.cpp
@@ -1218,11 +1218,11 @@ void OptimizerEngineStageData::setKSMaxRadiusScale(const double& aInput)
     mKSMaxRadiusScale = aInput;
 }
 
-double OptimizerEngineStageData::getProblemUpdateFrequency() const
+size_t OptimizerEngineStageData::getProblemUpdateFrequency() const
 {
     return mProblemUpdateFrequency;
 }
-void OptimizerEngineStageData::setProblemUpdateFrequency(const double& aInput)
+void OptimizerEngineStageData::setProblemUpdateFrequency(const size_t& aInput)
 {
     mProblemUpdateFrequency = aInput;
 }
